A static_cast necessary
git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@201 f17a810d-92cc-4fca-aeb8-6e32867107aa

+++ itpp/base/random.cpp
@@ -69,7 +69,7 @@ namespace itpp {
 
   void Random_Generator::randomize()
   {
-    lastSeed = time(0) * getpid(); // not good enough if randomize is used within a short time-interval
+    lastSeed = static_cast<unsigned long>(time(0)) * getpid(); // not good enough if randomize is used within a short time-interval
     reset();
   }
